+++ lib/registration_office/demand.rb
@@ -6,56 +6,53 @@ module RegistrationOffice
   # Instead, the +.[]+ singleton method should be used,
   # which returns a new module that can be included.
   module Demand
-    class << self
-      # @param registry_object [Object] The Ruby object (class, module) that defined the register which should be used.
-      #
-      # @return [Module]
-      def [](registry_object)
-        safe_constantize(registry_object)
+    class DuplicateDemandNameError < StandardError; end
 
-        mod = Module.new
-        prepare_mod(mod, registry_object)
-        mod
+    class << self
+      # rubocop:disable-next Metrics/MethodLength
+      def included(base)
+        mod = prepare_mod(Module.new)
+        base.module_eval do
+          const_set('RegistryDemand', mod)
+
+          define_singleton_method(:add_demand) do |register_name, register_object|
+            const_get('RegistryDemand').add_demand(
+              register_name,
+              register_object.registry(register_name)
+            )
+          end
+
+          define_singleton_method(:demand) { |register_name| const_get('RegistryDemand').demand(register_name) }
+          define_method(:demand) { |register_name| self.class.const_get('RegistryDemand').demand(register_name) }
+        end
       end
 
       private
 
-      def safe_constantize(registry_object)
-        return registry_object if registry_object.is_a?(Class) || registry_object.is_a?(Module)
-
-        unless registry_object.is_a?(String)
-          raise ArgumentError, "`#{registry_object.inspect}` must be a String or a Class or a Module"
-        end
-
-        Object.const_get(registry_object)
-      rescue NameError
-        raise ArgumentError, "`#{registry_object}` is not a known Class or Module"
-      end
-
-      # rubocop:disable-next Metrics/AbcSize, Metrics/MethodLength
-      def prepare_mod(mod, registry_object)
-        unless registry_object.respond_to?(:registry)
-          raise ArgumentError, "`#{registry_object.inspect}` does not respond to #registry"
-        end
-
-        # This is evaluated withing the dynamically created, nested module "RegistryDemand"
+      # rubocop:disable-next Metrics/MethodLength
+      def prepare_mod(mod)
         mod.module_eval do
-          # Eq. to `def self.included(base)`
-          define_singleton_method :included do |base|
-            # This is evaluated withing the including class/module
-            base.module_eval do
-              const_set('RegistryDemand', mod)
+          class << self
+            def demands
+              @demands ||= Registers.new
+            end
 
-              define_singleton_method(:demand) { const_get('RegistryDemand') }
-              define_method(:demand) { self.class.const_get('RegistryDemand') }
+            def demand(register_name)
+              demands.fetch(register_name)
+            end
+
+            def add_demand(register_name, register)
+              demands.register_a_register(register_name, register)
+            rescue Registers::DuplicateRegisterNameError
+              raise(
+                DuplicateDemandNameError,
+                "A register with name `#{register_name}` is already demanded"
+              )
             end
           end
-
-          # "Delegated" methods
-          define_singleton_method(:key!) { |key| registry_object.registry.key!(key) }
-          define_singleton_method(:use!) { |key| registry_object.registry.use!(key) }
-          define_singleton_method(:keys) { registry_object.registry.keys }
         end
+
+        mod
       end
     end
   end