+++ spec/features/abstract_full_integration_spec.rb
@@ -7,18 +7,22 @@ RSpec.describe 'abstract full integration' do
         include RegistrationOffice[:registration]
 
         register(
-          :invalid_bicycle_configuration,
-          :invalid_coupon_code,
-          :bicycle_not_in_stock,
-          :customer_not_solvent
+          :failures,
+          keys:
+            [
+              :invalid_bicycle_configuration,
+              :invalid_coupon_code,
+              :bicycle_not_in_stock,
+              :customer_not_solvent
+            ]
         )
 
         def self.self_demand_key(key)
-          demand.key!(key)
+          demand(:failures).key!(key)
         end
 
         def demand_key(key)
-          self.class.demand.key!(key)
+          self.class.demand(:failures).key!(key)
         end
       end
 
@@ -26,14 +30,16 @@ RSpec.describe 'abstract full integration' do
 
     order_class =
       Class.new do
-        include RegistrationOffice[:demand, registry_object: Dealer]
+        include RegistrationOffice[:demand]
+
+        add_demand(:failures, Dealer)
 
         def self.self_demand_key(key)
-          demand.key!(key)
+          demand(:failures).key!(key)
         end
 
         def demand_key(key)
-          demand.key!(key)
+          demand(:failures).key!(key)
         end
       end
 
@@ -54,7 +60,7 @@ RSpec.describe 'abstract full integration' do
     describe '.registry.all' do
       it do
         all_keys = [:invalid_bicycle_configuration, :invalid_coupon_code, :bicycle_not_in_stock, :customer_not_solvent]
-        expect(Dealer.registry.keys).to eq all_keys
+        expect(Dealer.registry(:failures).keys).to eq all_keys
       end
     end
   end
@@ -71,12 +77,14 @@ RSpec.describe 'abstract full integration' do
     end
 
     describe 'demanding an invalid key' do
+      let(:unregistered_key_error) { [RegistrationOffice::Register::UnregisteredKeyError, 'Register `failures` in `Dealer`: key `xxx` is not registered'] }
+
       context 'when called on instance' do
-        it { expect { Dealership::Order.new.demand_key(:xxx) }.to raise_error Dealer::RegistryStore::UnregisteredKeyError, 'key `xxx` is not registered' }
+        it { expect { Dealership::Order.new.demand_key(:xxx) }.to raise_error(*unregistered_key_error) }
       end
 
       context 'when called on class' do
-        it { expect { Dealership::Order.self_demand_key(:xxx) }.to raise_error Dealer::RegistryStore::UnregisteredKeyError, 'key `xxx` is not registered' }
+        it { expect { Dealership::Order.self_demand_key(:xxx) }.to raise_error(*unregistered_key_error) }
       end
     end
   end