+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## [Unreleased]
 
-- Enforce registering keys all at once:
-  Use `register(:key_one, :key_two, :key_three)` instead of multiple `register` calls.
+- Uses named registers and demands instead of anonymous
+- Allow multiple named registers and demands
+- Enforces registering keys of same register all at once
 
 ## [0.1.0] - 2026-09-04